create L1THGCalUtilities and move ntuples
