+++ L1Trigger/L1THGCal/plugins/BuildFile.xml
@@ -29,18 +29,3 @@
   <flags   EDM_PLUGIN="1"/>
 </library>
 
-<library   name="L1TriggerL1THGCalPlugins_ntuples" file="ntuples/*.cc">
-  <use   name="L1Trigger/L1THGCal"/>
-  <use   name="CommonTools/UtilAlgos"/>
-  <use   name="SimDataFormats/CaloTest"/>
-  <use   name="DataFormats/JetReco"/>
-  <use   name="Geometry/HcalTowerAlgo"/>
-  <use   name="SimDataFormats/PileupSummaryInfo"/>
-  <use   name="SimDataFormats/GeneratorProducts"/>
-  <use   name="TrackPropagation/RungeKutta"/>
-  <use   name="FastSimulation/Event"/>
-  <use   name="MagneticField/Engine"/>
-  <use   name="MagneticField/Records"/>
-  <use name="heppdt"/>
-  <flags   EDM_PLUGIN="1"/>
-</library>
