+++ L1Trigger/L1THGCal/test/testHGCalL1T_RelVal_cfg.py
@@ -54,11 +54,15 @@ process.configurationMetadata = cms.untracked.PSet(
     name = cms.untracked.string('Applications')
 )
 
-# Output definition
-process.TFileService = cms.Service(
-    "TFileService",
-    fileName = cms.string("ntuple.root")
-    )
+process.FEVTDEBUGoutput = cms.OutputModule("PoolOutputModule",
+    splitLevel = cms.untracked.int32(0),
+    eventAutoFlushCompressedSize = cms.untracked.int32(5242880),
+    outputCommands = cms.untracked.vstring(
+        'keep *_hgcalBackEndLayer2Producer_*_*',
+        'keep *_hgcalTowerProducer_*_*',
+    ),
+    fileName = cms.untracked.string('file:test.root')
+)
 
 # Other statements
 from Configuration.AlCa.GlobalTag import GlobalTag
@@ -71,12 +75,11 @@ process.hgcl1tpg_step = cms.Path(process.hgcalTriggerPrimitives)
 #  from L1Trigger.L1THGCal.customTriggerGeometry import custom_geometry_ZoltanSplit_V7
 #  process = custom_geometry_ZoltanSplit_V7(process)
 
-# load ntuplizer
-process.load('L1Trigger.L1THGCal.hgcalTriggerNtuples_cff')
-process.ntuple_step = cms.Path(process.hgcalTriggerNtuples)
+
+process.FEVTDEBUGoutput_step = cms.EndPath(process.FEVTDEBUGoutput)
 
 # Schedule definition
-process.schedule = cms.Schedule(process.hgcl1tpg_step, process.ntuple_step)
+process.schedule = cms.Schedule(process.hgcl1tpg_step, process.FEVTDEBUGoutput_step)
 
 # Add early deletion of temporary data products to reduce peak memory need
 from Configuration.StandardSequences.earlyDeleteSettings_cff import customiseEarlyDelete
