+++ L1Trigger/L1THGCalUtilities/plugins/BuildFile.xml
@@ -1,0 +1,17 @@
+<library   name="L1TriggerL1THGCalUtilitiesPlugins_ntuples" file="ntuples/*.cc">
+  <use   name="L1Trigger/L1THGCal"/>
+  <use   name="L1Trigger/L1THGCalUtilities"/>
+  <use   name="CommonTools/UtilAlgos"/>
+  <use   name="SimDataFormats/CaloTest"/>
+  <use   name="DataFormats/JetReco"/>
+  <use   name="Geometry/HcalTowerAlgo"/>
+  <use   name="SimDataFormats/PileupSummaryInfo"/>
+  <use   name="SimDataFormats/GeneratorProducts"/>
+  <use   name="TrackPropagation/RungeKutta"/>
+  <use   name="FastSimulation/Event"/>
+  <use   name="FastSimulation/CaloGeometryTools"/>
+  <use   name="MagneticField/Engine"/>
+  <use   name="MagneticField/Records"/>
+  <use name="heppdt"/>
+  <flags   EDM_PLUGIN="1"/>
+</library>
