+++ L1Trigger/L1THGCalUtilities/plugins/ntuples/HGCalTriggerNtupleEvent.cc
@@ -1,4 +1,4 @@
-#include "L1Trigger/L1THGCal/interface/HGCalTriggerNtupleBase.h"
+#include "L1Trigger/L1THGCalUtilities/interface/HGCalTriggerNtupleBase.h"
 
 class HGCalTriggerNtupleEvent : public HGCalTriggerNtupleBase
 {
