+++ L1Trigger/L1THGCalUtilities/plugins/ntuples/HGCalTriggerNtupleGen.cc
@@ -4,7 +4,7 @@
 #include "DataFormats/HcalDetId/interface/HcalDetId.h"
 
 
-#include "L1Trigger/L1THGCal/interface/HGCalTriggerNtupleBase.h"
+#include "L1Trigger/L1THGCalUtilities/interface/HGCalTriggerNtupleBase.h"
 #include "L1Trigger/L1THGCal/interface/HGCalTriggerTools.h"
 
 #include "MagneticField/Engine/interface/MagneticField.h"
