+++ L1Trigger/L1THGCalUtilities/src/HGCalTriggerNtupleBase.cc
@@ -1,4 +1,4 @@
-#include "L1Trigger/L1THGCal/interface/HGCalTriggerNtupleBase.h"
+#include "L1Trigger/L1THGCalUtilities/interface/HGCalTriggerNtupleBase.h"
 
 EDM_REGISTER_PLUGINFACTORY(HGCalTriggerNtupleFactory,
                            "HGCalTriggerNtupleFactory");
